Refactor: main.js에서 post 리퀘스트 처리 post.js로 분리
로그인 버튼 클릭 했을 때 발생하는 post request를 post.js로 분리했습니다.
get방식은 main.js에 post방식은 post.js에 추가하도록 하겠습니다.
